% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionModel.R
\name{LR.model}
\alias{LR.model}
\title{Logistic regression model}
\usage{
LR.model(train.data, test.data)
}
\arguments{
\item{train.data}{[data.frame] The training dataset. It must contain a column named Y providing the case/control phenotype (0 = unaffected (control), 1 = affected (case)).}

\item{test.data}{[data.frame] The test dataset. It must contain a column named Y providing the case/control phenotype (0 = unaffected (control), 1 = affected (case)).}
}
\value{
\code{LR.model} return a vector containing the predicted probability.
}
\description{
A Function to build a logistic regression prediction model on the training dataset, and make predictions on the test dataset.
}
\examples{
input.dir <- system.file("extdata", package="PrismVote")
output.dir <- system.file("extdata", package="PrismVote")
path2plink <- '/path/to/plink'
\dontrun{
covar.number <-  c(2, 3)

feature.selection.result <- feature.selection(input.dir = input.dir,
output.dir = output.dir,
genotype = "train",
phenotype = "train.phenotypes.txt",
covar.number = covar.number,
plink.path = path2plink,
topK = 10,
verbose = TRUE)

train.genotype.path <- file.path(input.dir, "train.raw")
train.genotype.path <- gsub('\\\\\\\\', '/', train.genotype.path)

train.phenotype.path <- file.path(input.dir, "train.phenotypes.txt")
train.phenotype.path <- gsub('\\\\\\\\', '/', train.phenotype.path)

train.data <- data.table::fread(train.genotype.path, data.table = FALSE)[, -c(1,2,3,4,5,6)]
colnames(train.data) <- unlist(purrr::map(colnames(train.data),function(x) {substr(x,1, nchar(x)-2)}))
train.pheno <- data.table::fread(train.phenotype.path, data.table = FALSE)
train.data$Y <- train.pheno[, 3]-1
train.data <- cbind(train.data[, feature.selection.result$index, drop = FALSE],
train.pheno[, covar.number + 2, drop = FALSE])

test.genotype.path <- file.path(input.dir, "test.raw")
test.genotype.path <- gsub('\\\\\\\\', '/', test.genotype.path)

test.phenotype.path <- file.path(input.dir, "test.phenotypes.txt")
test.phenotype.path <- gsub('\\\\\\\\', '/', test.phenotype.path)

test.data <- data.table::fread(test.genotype.path, data.table = FALSE)[, -c(1,2,3,4,5,6)]
colnames(test.data) <- unlist(purrr::map(colnames(test.data),function(x) {substr(x,1, nchar(x)-2)}))
test.pheno <- data.table::fread(test.phenotype.path, data.table = FALSE)
test.data$Y <- test.pheno[, 3]-1
test.data <- cbind(test.data[, feature.selection.result$index, drop = FALSE],
test.pheno[, covar.number + 2, drop = FALSE])

LR.pred <- LR.model(train.data, test.data)
}
}
