% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Principal component analysis}
\usage{
PCA(
  input.dir,
  output.dir,
  train.genotype,
  test.genotype,
  PCA.separate = FALSE,
  PCs.count = 10,
  plink.path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input.dir}{[character] The full absolute path to the directory containing the training and test dataset. If \code{input.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{output.dir}{[character] The full absolute path where the result will be written to. If \code{output.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{train.genotype}{[character] The prefix of PLINK binary files (bed/bim/fam) of the training dataset.}

\item{test.genotype}{[character] The prefix of PLINK binary files (bed/bim/fam) of the test dataset.}

\item{PCA.separate}{[logical] If TURE, the principal components are calculated from the training dataset and then project the test dataset onto those principal components. If FALSE, the principal components are calculated from the combined data of the training and test dataset. The default value is FALSE.}

\item{PCs.count}{[numeric] To specify the number of top principal components that should be extracted. The default value is 10.}

\item{plink.path}{[character] The full absolute path to the PLINK executable file. The executable to run is path/to/plink.exe if you are on Windows operating system, for Unix-like operating system this is path/to/plink. If \code{plink.path} is NULL, the PLINK PATH should be added as a system environment variable.}

\item{verbose}{[logical] If TRUE, the PLINK log, error, and warning information are printed to standard out. The default value is TRUE.}
}
\value{
\code{PCA} returns a list containing eigenvalues and eigenvectors of the training and test dataset:
\item{eigenvalue}{A vector containing the top eigenvalues according to \code{PCs.count} specified.}
\item{train.eigenvector}{A data frame containing the eigenvectors of the training dataset.}
\item{test.eigenvector}{A data frame containing the eigenvectors of the test dataset.}
}
\description{
A Function to run principal component analysis on input dataset using PLINK 1.9.
}
\examples{
input.dir <- system.file("extdata", package="pv")
output.dir <- system.file("extdata", package="pv")
path2plink <- '/path/to/plink'
\dontrun{
pca.result <- PCA(input.dir = input.dir,
output.dir = output.dir,
train.genotype = "train",
test.genotype = "test",
PCA.separate = FALSE,
PCs.count = 10,
plink.path = path2plink,
verbose = TRUE)
}
}
