% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{plink.lr}
\alias{plink.lr}
\title{Logistic regression analysis}
\usage{
plink.lr(
  input.dir,
  output.dir,
  genotype,
  phenotype = NULL,
  covar.number = NULL,
  plink.path = NULL,
  verbose = TRUE
)
}
\arguments{
\item{input.dir}{[character] The full absolute path to the directory containing the training and test dataset. If \code{input.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{output.dir}{[character] The full absolute path where the result will be written to. If \code{output.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{genotype}{[character] The prefix of PLINK binary files (bed/bim/fam).}

\item{phenotype}{[character] To specify an alternate phenotype for the logistic regression analysis using the "\code{--pheno}" flag in plink. This file must have a header row. The first and second columns of the phenotype file must be "FID" and "IID", the case/control phenotype in column 3 (1 = unaffected (control), 2 = affected (case)), and covariates in remaining columns. If \code{phenotype} is NULL, the original fam file must contain a phenotype in column 6. See the PLINK 1.9 documentation for details (\url{https://www.cog-genomics.org/plink/1.9/}).}

\item{covar.number}{[vector] To specify a subset of column numbers of covariates to load from \code{phenotype} file using the "\code{--covar-number}" flag in plink. If NULL (the default), the logistic regression model without covariate adjustment.}

\item{plink.path}{[character] The full absolute path to the PLINK executable file. The executable to run is path/to/plink.exe if you are on a Windows operating system, for Unix-like operating system this is path/to/plink. If \code{plink.path} is NULL, the PLINK PATH should be added as a system environment variable.}

\item{verbose}{[logical] If TRUE, the PLINK log, error, and warning information are printed to standard out. The default value is TRUE.}
}
\value{
\code{plink.lr} returns a data frame with the results of logistic regression analysis derived from PLINK.
}
\description{
A Function to perform logistic regression analysis given a case/control phenotype using PLINK 1.9.
}
\examples{
input.dir <- system.file("extdata", package="pv")
output.dir <- system.file("extdata", package="pv")
path2plink <- '/path/to/plink'
\dontrun{
lr.result <- plink.lr(input.dir = input.dir,
output.dir = output.dir,
genotype = "train",
phenotype = "train.phenotypes.txt",
covar.number = c(2, 3),
plink.path = path2plink,
verbose = TRUE)
}
}
