% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrismVote.R
\name{PV}
\alias{PV}
\title{Prism Vote}
\usage{
PV(
  input.dir,
  output.dir,
  train.genotype,
  train.phenotype,
  test.genotype,
  test.phenotype,
  covar.number.PV = NULL,
  covar.number.LR = NULL,
  PCA.separate = FALSE,
  PCs.count = 10,
  stratum.count = 2,
  plink.path = NULL,
  P.value = NULL,
  topK = 10,
  candidate.SNPs = NULL,
  CS = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input.dir}{[character] The full absolute path to the directory containing the training and test dataset. If \code{input.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{output.dir}{[character] The full absolute path where the result will be written to. If \code{output.dir} is missing, the current working directory obtained by \code{getwd()} is used.}

\item{train.genotype}{[character] The prefix of PLINK binary files (bed/bim/fam) of the training dataset.}

\item{train.phenotype}{[character] A space- or tab-delimited file to specify an alternate phenotype of the training dataset for the logistic regression analysis using the "\code{--pheno}" flag in plink. This file must have a header row. The first and second columns of the phenotype file must be "FID" and "IID", the case/control phenotype in column 3 (1 = unaffected (control), 2 = affected (case)), and covariates in remaining columns. See the PLINK 1.9 documentation for details (\url{https://www.cog-genomics.org/plink/1.9/}).}

\item{test.genotype}{[character] The prefix of PLINK binary files (bed/bim/fam) of the test dataset.}

\item{test.phenotype}{[character] A space- or tab-delimited file to specify an alternate phenotype of the training dataset for the logistic regression analysis using the "\code{--pheno}" flag in plink. This file must have a header row. The first and second columns of the phenotype file must be "FID" and "IID", the case/control phenotype in column 3 (1 = unaffected (control), 2 = affected (case)), and covariates in remaining columns.  See the PLINK 1.9 documentation for details (\url{https://www.cog-genomics.org/plink/1.9/}).}

\item{covar.number.PV}{[vector] Used in PV model to specify a subset of column numbers of covariates to load from \code{phenotype} file using the "\code{--covar-number}" flag in plink (via \code{\link{plink.lr}}). If NULL (the default), the logistic regression model without covariate adjustment.}

\item{covar.number.LR}{[vector] Used in LR and PRS model to specify a subset of column numbers of covariates to load from \code{phenotype} file using the "\code{--covar-number}" flag in plink (via \code{\link{plink.lr}}). If NULL (the default), the logistic regression model without covariate adjustment.}

\item{PCA.separate}{[logical] If TURE, the principal components are calculated from the training dataset and then project the test dataset onto those principal components. If FALSE, the principal components are calculated from the combined data of the training and test dataset. The default value is FALSE.}

\item{PCs.count}{[numeric] To specify the number of top principal components that should be extracted. The default value is 10.}

\item{stratum.count}{[numeric] To specify the number of strata, as default the sample size of each stratum is N/stratum.count, N is the sample size. The default value is 2.}

\item{plink.path}{[character] The full absolute path to the PLINK executable file. The executable to run is path/to/plink.exe if you are on a Windows operating system, for Unix-like operating system this is path/to/plink. If \code{plink.path} is NULL, the PLINK PATH should be added as a system environment variable.}

\item{P.value}{[double] To specify the genome-wide significance P-value threshold to select the significant SNPs to build a prediction model. The default value is NULL. This value is ignored when candidate.SNPs is not NULL. When left NULL (the default), the topK or candidate.SNPs will be used. The P-value of each SNP is calculated from logistic regression analysis using PLINK 1.9 (via \code{\link{plink.lr}}).}

\item{topK}{[numeric] To specify the top K significant SNPs to build a prediction model. For a fair comparison, the number of the top-ranked SNPs from entire sample (for LR and PRS model) equals to the number of the unique union set of the selected SNPs from each stratum in PV. The default value is 10. This value is ignored when P.value or candidate.SNPs is not NULL.}

\item{candidate.SNPs}{[vector] A character vector of SNP name, used to specify the candidate SNPs to build a prediction model, ignores \code{P.value} and \code{topK}. The default value is NULL. Should match the names of SNPs in the provided PLINK binary files.}

\item{CS}{[logical] If TRUE, the softmax of cosine similarity will be used to calculate the probability that the samples belong to each stratum. If FALSE, the squared distance of a subject to a cluster center empirically follows a chi-squared distribution will be used. The default value is FALSE.}

\item{verbose}{[logical] If TRUE, the PLINK log, error, and warning information are printed to standard out. The default value is TRUE.}
}
\value{
\code{PV} returns a list with the following components:
\item{stratification.result}{The output of \code{\link{stratification}}}
\item{feature.selection.result}{The output of \code{\link{feature.selection}}}
\item{LR.result}{The results of logistic regression model. A list with i) predict, the output of \code{\link{LR.model}}. ii) performance, a list containing the AUC and accuracy (Acc) value.}
\item{LR.PV.result}{The results of the logistic regression model under the Prism Vote framework. A list with i) predict, the output of \code{\link{LR.PV.model}}. ii) performance, a list containing the AUC and accuracy (Acc) value.}
\item{PRS.result}{The results of the logistic regression model based on the polygenic risk score. A list with i).predict, the output of \code{\link{LR.model}}. ii) performance, a list containing the AUC and accuracy (Acc) value.}
\item{PRS.PV.result}{The results of the logistic regression model based on the polygenic risk score under the Prism Vote framework. A list with i) predict, the output of \code{\link{LR.PV.model}}. ii) performance, a list containing the AUC and accuracy (Acc) value.}
}
\description{
Perform Prism Vote method to build a prediction model on the training dataset, and make predictions on the test dataset.
}
\examples{
input.dir <- system.file("extdata", package="pv")
output.dir <- system.file("extdata", package="pv")
path2plink <- '/path/to/plink'
\dontrun{
pv.result <- PV(input.dir = input.dir,
output.dir = output.dir,
train.genotype = "train",
train.phenotype = "train.phenotypes.txt",
test.genotype = "test",
test.phenotype = "test.phenotypes.txt",
covar.number.PV = c(2,3),
covar.number.LR = c(2,3),
PCA.separate = FALSE,
PCs.count = 10,
stratum.count = 2,
plink.path = path2plink,
P.value = NULL,
topK = 10,
CS = FALSE,
candidate.SNPs = NULL,
verbose = TRUE)
}
}
\seealso{
{\code{\link{PCA}}, \code{\link{stratification}}, \code{\link{LR.model}}, \code{\link{PV.model}}, \code{\link{PRS}}, \code{\link{feature.selection}}}
}
